refactor(updater): 使用flock替代文件锁原生实现
